Adding option to run tests with no-race check
