+++ pkg/ingester/ingester_test.go
@@ -5273,69 +5273,6 @@ func TestExpendedPostingsCacheIsolation(t *testing.T) {
 	wg.Wait()
 }
 
-func TestExpandedCachePostings_Race(t *testing.T) {
-	cfg := defaultIngesterTestConfig(t)
-	cfg.BlocksStorageConfig.TSDB.BlockRanges = []time.Duration{2 * time.Hour}
-	cfg.LifecyclerConfig.JoinAfter = 0
-	cfg.BlocksStorageConfig.TSDB.PostingsCache.Head.Enabled = true
-
-	r := prometheus.NewRegistry()
-	i, err := prepareIngesterWithBlocksStorage(t, cfg, r)
-	require.NoError(t, err)
-	require.NoError(t, services.StartAndAwaitRunning(context.Background(), i))
-	defer services.StopAndAwaitTerminated(context.Background(), i) //nolint:errcheck
-
-	// Wait until the ingester is ACTIVE
-	test.Poll(t, 100*time.Millisecond, ring.ACTIVE, func() interface{} {
-		return i.lifecycler.GetState()
-	})
-
-	ctx := user.InjectOrgID(context.Background(), "test")
-
-	wg := sync.WaitGroup{}
-
-	for j := 0; j < 100; j++ {
-		metricName := fmt.Sprintf("test_metric_%d", j)
-		wg.Add(200)
-		for k := 0; k < 100; k++ {
-
-			go func() {
-				defer wg.Done()
-				_, err := i.Push(ctx, cortexpb.ToWriteRequest(
-					[]labels.Labels{labels.FromStrings(labels.MetricName, metricName, "k", strconv.Itoa(k))},
-					[]cortexpb.Sample{{Value: 1, TimestampMs: 9}}, nil, nil, cortexpb.API))
-				require.NoError(t, err)
-			}()
-
-			go func() {
-				defer wg.Done()
-				err := i.QueryStream(&client.QueryRequest{
-					StartTimestampMs: 0,
-					EndTimestampMs:   math.MaxInt64,
-					Matchers:         []*client.LabelMatcher{{Type: client.EQUAL, Name: labels.MetricName, Value: metricName}},
-				}, &mockQueryStreamServer{ctx: ctx})
-				require.NoError(t, err)
-			}()
-		}
-
-		wg.Wait()
-
-		s := &mockQueryStreamServer{ctx: ctx}
-		err = i.QueryStream(&client.QueryRequest{
-			StartTimestampMs: 0,
-			EndTimestampMs:   math.MaxInt64,
-			Matchers:         []*client.LabelMatcher{{Type: client.EQUAL, Name: labels.MetricName, Value: metricName}},
-		}, s)
-		require.NoError(t, err)
-
-		set, err := seriesSetFromResponseStream(s)
-		require.NoError(t, err)
-		res, err := client.MatrixFromSeriesSet(set)
-		require.NoError(t, err)
-		require.Equal(t, 100, res.Len())
-	}
-}
-
 func TestExpendedPostingsCache(t *testing.T) {
 	cfg := defaultIngesterTestConfig(t)
 	cfg.BlocksStorageConfig.TSDB.BlockRanges = []time.Duration{2 * time.Hour}
